refactor: refactoring orders-page[orders]
