+++ src/components/Pages/index.ts
@@ -1,3 +1,4 @@
 export { Clients } from "./Clients/Clients";
 export { Category } from "./Category/Category";
 export { Brands } from "./Brands/Brands";
+export { Orders } from "./Orders/Orders";
